Homework 9/local_version/requirements.txt
numpy==1.19.5
Pillow==8.4.0
tensorflow==2.7.0
